+++ Angular-Front-End/app-front-end/src/app/models/community.model.ts
@@ -1,5 +1,7 @@
 import { Identifiable } from './identifiable';
 
 export interface CommunityModel extends Identifiable{
-    name: String;
+    name: string;
+    description: string;
+    menuShow?: boolean;
 }